backend/requirements.txt
django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
django-redis==6.0.0
psycopg[binary]==3.3.2
sentry-sdk==2.52.0
gunicorn==23.0.0
